fix: tensor type

+++ lymph_nodes/data/datasets/segmentation_dataset.py
@@ -5,6 +5,7 @@ from typing import Any, TypeVar
 
 import mlflow
 import numpy as np
+import torch
 from numpy.typing import NDArray
 from openslide import OpenSlide
 from rationai.masks import (
@@ -108,7 +109,7 @@ class _SegmentationSlideTiles(MaskSlideTiles[Sample]):
         image = tensor["image"]
         mask = tensor["mask"]
 
-        return image, mask, metadata
+        return image, mask.to(torch.uint8), metadata
 
 
 class _SegmentationSlideTilesPred(MaskSlideTiles[PredictSample]):
